test(scheduler): split reclaim topology tests from benchmarks
